requirements.txt
# "fastapi[standard]"
sentence-transformers
# pydantic
# uvicorn
pymysql
sqlmodel
# python-dotenv
pydantic-settings
mysqlclient